Adding the space for Consent Form

+++ report/AccountLedgerReport/accountLedger.js
@@ -107,7 +107,7 @@ function populateTable4(data) {
           cell.textContent = utcDate.toLocaleString("en-IN", options);
         } else if (key === "online_amt") {
           cell.textContent = item[key];
-          grandTotalQuantity += parseInt(item[key]);
+          grandTotalQuantity += parseFloat(item[key]);
         } else {
           cell.textContent = item[key];
         }
